+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fix: Panic during crate installation in case args are empty #615
 * Enhancement: New list-category-steps command #603 (thanks @grbd)
 * Enhancement: New tls feature so tls can be disabled (by default enabled) #614
+* Fix: If rust-script is used and requires installation, install old 0.7.0 version due to rust-script bug
 
 ### v0.35.6 (2021-11-01)